Align feature class naming to a new scheme.
