+++ src/headingui.js
@@ -4,38 +4,124 @@
  */
 
 /**
- * @module heading/heading
+ * @module heading/headingui
  */
 
-import HeadingEditing from './headingediting';
-import HeadingUI from './headingui';
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
+import Model from '@ckeditor/ckeditor5-ui/src/model';
+
+import { createDropdown, addListToDropdown } from '@ckeditor/ckeditor5-ui/src/dropdown/utils';
+
+import Collection from '@ckeditor/ckeditor5-utils/src/collection';
 
 import '../theme/heading.css';
 
 /**
- * The headings feature.
- *
- * It loads the {@link module:heading/headingediting~HeadingEditing heading editing feature}
- * and {@link module:heading/headingui~HeadingUI heading UI feature}.
- *
- * For a detailed overview, check the {@glink features/headings Headings feature documentation}.
+ * The headings UI feature. It introduces the `headings` drop-down.
  *
  * @extends module:core/plugin~Plugin
  */
-export default class Heading extends Plugin {
+export default class HeadingUI extends Plugin {
 	/**
 	 * @inheritDoc
 	 */
-	static get requires() {
-		return [ HeadingEditing, HeadingUI ];
+	init() {
+		const editor = this.editor;
+		const t = editor.t;
+		const options = this._getLocalizedOptions();
+		const defaultTitle = t( 'Choose heading' );
+		const dropdownTooltip = t( 'Heading' );
+
+		// Register UI component.
+		editor.ui.componentFactory.add( 'headings', locale => {
+			const commands = [];
+			const dropdownItems = new Collection();
+
+			for ( const option of options ) {
+				const command = editor.commands.get( option.model );
+				const itemModel = new Model( {
+					commandName: option.model,
+					label: option.title,
+					class: option.class
+				} );
+
+				itemModel.bind( 'isActive' ).to( command, 'value' );
+
+				// Add the option to the collection.
+				dropdownItems.add( itemModel );
+
+				commands.push( command );
+			}
+
+			const dropdownView = createDropdown( locale );
+			addListToDropdown( dropdownView, dropdownItems );
+
+			dropdownView.buttonView.set( {
+				isOn: false,
+				withText: true,
+				tooltip: dropdownTooltip
+			} );
+
+			dropdownView.extendTemplate( {
+				attributes: {
+					class: [
+						'ck-heading-dropdown'
+					]
+				}
+			} );
+
+			dropdownView.bind( 'isEnabled' ).toMany( commands, 'isEnabled', ( ...areEnabled ) => {
+				return areEnabled.some( isEnabled => isEnabled );
+			} );
+
+			dropdownView.buttonView.bind( 'label' ).toMany( commands, 'value', ( ...areActive ) => {
+				const index = areActive.findIndex( value => value );
+
+				// If none of the commands is active, display default title.
+				return options[ index ] ? options[ index ].title : defaultTitle;
+			} );
+
+			// Execute command when an item from the dropdown is selected.
+			this.listenTo( dropdownView, 'execute', evt => {
+				editor.execute( evt.source.commandName );
+				editor.editing.view.focus();
+			} );
+
+			return dropdownView;
+		} );
 	}
 
 	/**
-	 * @inheritDoc
+	 * Returns heading options as defined in `config.heading.options` but processed to consider
+	 * editor localization, i.e. to display {@link module:heading/heading~HeadingOption}
+	 * in the correct language.
+	 *
+	 * Note: The reason behind this method is that there's no way to use {@link module:utils/locale~Locale#t}
+	 * when the user config is defined because the editor does not exist yet.
+	 *
+	 * @private
+	 * @returns {Array.<module:heading/heading~HeadingOption>}.
 	 */
-	static get pluginName() {
-		return 'Heading';
+	_getLocalizedOptions() {
+		const editor = this.editor;
+		const t = editor.t;
+		const localizedTitles = {
+			Paragraph: t( 'Paragraph' ),
+			'Heading 1': t( 'Heading 1' ),
+			'Heading 2': t( 'Heading 2' ),
+			'Heading 3': t( 'Heading 3' )
+		};
+
+		return editor.config.get( 'heading.options' ).map( option => {
+			const title = localizedTitles[ option.title ];
+
+			if ( title && title != option.title ) {
+				// Clone the option to avoid altering the original `config.heading.options`.
+				option = Object.assign( {}, option, { title } );
+			}
+
+			return option;
+		} );
 	}
 }
 
@@ -106,9 +192,11 @@ export default class Heading extends Plugin {
 /**
  * Heading option descriptor.
  *
+ * This format is compatible with {@link module:engine/conversion/definition-based-converters~ConverterDefinition}
+ * and adds to additional properties: `title` and `class`.
+ *
  * @typedef {Object} module:heading/heading~HeadingOption
- * @property {String} model Name of the model element to convert.
- * @property {module:engine/view/elementdefinition~ElementDefinition} view Definition of a view element to convert from/to.
+ * @extends module:engine/conversion/definition-based-converters~ConverterDefinition
  * @property {String} title The user-readable title of the option.
  * @property {String} class The class which will be added to the dropdown item representing this option.
  */
